DataLoader to DataLoaderConfig
