+++ clinicadl/data/dataloader/__init__.py
@@ -3,4 +3,4 @@ adapted to ``ClinicaDL``."""
 
 from .batch import Batch, BatchType
 from .collate import *
-from .config import DataLoader, DataLoaderConfig
+from .loader import DataLoader
